requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
apscheduler==3.11.0
python-dotenv==1.0.1
jinja2==3.1.5
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
chinese-calendar==1.9.2
